pyproject.toml
[tool.poetry]
name = "The Discord Archivist"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.7.4"
async-timeout = "^3.0.1"
attrs = "^21.2.0"
autopep8 = "^1.5.7"
certifi = "^2020.12.5"
chardet = "^4.0.0"
"discord.py" = "^1.7.2"
EbookLib = "^0.17.1"
idna = "^3.1"
lxml = "^4.6.3"
multidict = "^5.1.0"
Pillow = "^8.2.0"
pip = "^21.1.2"
pip-autoremove = "^0.9.1"
pycodestyle = "^2.7.0"
python-dateutil = "^2.8.1"
python-dotenv = "^0.17.1"
setuptools = "^57.0.0"
six = "^1.16.0"
toml = "^0.10.2"
typing-extensions = "^3.10.0"
wheel = "^0.36.2"
yarl = "^1.6.3"
Flask = "^2.0.1"
waitress = "^2.0.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
